1660 - add VictoryCities chart
