+++ game-core/src/main/java/games/strategy/engine/stats/Statistics.java
@@ -15,4 +15,5 @@ public class Statistics {
   private final Table<String, Round, Double> productionOfPlayerInRound = HashBasedTable.create();
   private final Table<String, Round, Double> tuvOfPlayerInRound = HashBasedTable.create();
   private final Table<String, Round, Double> unitsOfPlayerInRound = HashBasedTable.create();
+  private final Table<String, Round, Double> victoryCitiesOfPlayerInRound = HashBasedTable.create();
 }
